+++ src/components/QueryControls/QueryControls.js
@@ -1,0 +1,20 @@
+import React from 'react'
+
+import ReactAux from '../../hoc/ReactAux/ReactAux'
+import QueryControl from './QueryControl/QueryControl'
+
+const QueryControls = (props) => (
+    <ReactAux>
+        <QueryControl/>
+        <QueryControl/>
+        <QueryControl/>
+        <QueryControl/>
+        <QueryControl/>
+        <QueryControl/>
+        <QueryControl/>
+        <QueryControl/>
+    </ReactAux>
+
+)
+
+export default QueryControls
