--- src/components/QueryInterface/QueryControls/QueryControls.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-import ReactAux from '../../../hoc/ReactAux'
-import QueryControl from './QueryControl/QueryControl'
-
-const QueryControls = (props) => (
-    <ReactAux>
-        <QueryControl/>
-        <QueryControl/>
-        <QueryControl/>
-        <QueryControl/>
-        <QueryControl/>
-        <QueryControl/>
-        <QueryControl/>
-        <QueryControl/>
-    </ReactAux>
-
-)
-
-export default QueryControls
