--- src/components/QueryInterface/QueryResults/QueryResult/QueryResult.css
@@ -1,3 +1,0 @@
-.layout{
-    margin: 25px 50px 75px 50px;
-}
