--- src/components/QueryInterface/QueryResults/QueryResults.js
@@ -1,25 +1,0 @@
-import React from 'react'
-
-import ReactAux from '../../../hoc/ReactAux'
-import QueryResult from './QueryResult/QueryResult'
-
-const QueryResults = (props) => (
-    <ReactAux>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-        <QueryResult/>
-    </ReactAux>
-
-)
-
-export default QueryResults
