+++ src/components/QueryResults/QueryResult/QueryResult.css
@@ -1,0 +1,3 @@
+.layout{
+    margin: 25px 50px 75px 50px;
+}
