+++ src/components/QueryResults/QueryResult/QueryResult.js
@@ -1,0 +1,42 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { withStyles } from 'material-ui/styles';
+import ExpansionPanel, {
+    ExpansionPanelSummary,
+    ExpansionPanelDetails,
+} from 'material-ui/ExpansionPanel';
+import Typography from 'material-ui/Typography';
+import ExpandMoreIcon from 'material-ui-icons/ExpandMore';
+
+import classes from './QueryResult.css'
+
+const styles = theme => ({
+    root: {
+        width: '100%',
+    },
+    heading: {
+        fontSize: theme.typography.pxToRem(15),
+        fontWeight: theme.typography.fontWeightRegular,
+    },
+});
+
+function QueryResult(props) {
+    return (
+            <ExpansionPanel>
+                <ExpansionPanelSummary expandIcon={<ExpandMoreIcon />}>
+                    <Typography className={classes.heading}>Query Result</Typography>
+                </ExpansionPanelSummary>
+                <ExpansionPanelDetails>
+                    <Typography>
+                        Some more information.
+                    </Typography>
+                </ExpansionPanelDetails>
+            </ExpansionPanel>
+    );
+}
+
+QueryResult.propTypes = {
+    classes: PropTypes.object.isRequired,
+};
+
+export default withStyles(styles)(QueryResult);
