+++ src/components/QueryResults/QueryResults.js
@@ -1,0 +1,25 @@
+import React from 'react'
+
+import ReactAux from '../../hoc/ReactAux/ReactAux'
+import QueryResult from './QueryResult/QueryResult'
+
+const QueryResults = (props) => (
+    <ReactAux>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+        <QueryResult/>
+    </ReactAux>
+
+)
+
+export default QueryResults
