+++ src/containers/Header/Header.js
@@ -7,15 +7,16 @@ import Typography from 'material-ui/Typography';
 import IconButton from 'material-ui/IconButton';
 import AccountCircle from 'material-ui-icons/AccountCircle';
 import Switch from 'material-ui/Switch';
-import {FormControlLabel, FormGroup} from 'material-ui/Form';
+import {FormControlLabel} from 'material-ui/Form';
 import Menu, {MenuItem} from 'material-ui/Menu';
-import {Link} from 'react-router-dom'
+import {Link} from 'react-router-dom';
+import {connect} from 'react-redux';
 
 import Logo from '../../components/Logo/Logo'
 
 const styles = {
-    root: {
-        width: '100%',
+    appBar: {
+        backgroundColor: "transparent",
     },
     flex: {
         flex: 1,
@@ -28,12 +29,15 @@ const styles = {
 
 class Header extends React.Component {
     state = {
-        auth: true,
         anchorEl: null,
     };
 
     handleChange = (event, checked) => {
-        this.setState({auth: checked});
+        if (checked) {
+            return this.props.onLogin();
+        } else {
+            return this.props.onLogout();
+        }
     };
 
     handleMenu = event => {
@@ -46,57 +50,53 @@ class Header extends React.Component {
 
     render() {
         const {classes} = this.props;
-        const {auth, anchorEl} = this.state;
+        const {anchorEl} = this.state;
         const open = Boolean(anchorEl);
 
         return (
-            <div className={classes.root}>
-                <FormGroup>
-                </FormGroup>
-                <AppBar position="static">
-                    <Toolbar>
-                        <Link to="/"><Logo/></Link>
-                        <Typography type="title" color="inherit" className={classes.flex}>
-                            This is the header, you may modify it......
-                        </Typography>
-                        {auth && (
-                            <div>
-                                <IconButton
-                                    aria-owns={open ? 'menu-appbar' : null}
-                                    aria-haspopup="true"
-                                    onClick={this.handleMenu}
-                                    color="inherit"
-                                >
-                                    <AccountCircle/>
-                                </IconButton>
-                                <Menu
-                                    id="menu-appbar"
-                                    anchorEl={anchorEl}
-                                    anchorOrigin={{
-                                        vertical: 'top',
-                                        horizontal: 'right',
-                                    }}
-                                    transformOrigin={{
-                                        vertical: 'top',
-                                        horizontal: 'right',
-                                    }}
-                                    open={open}
-                                    onClose={this.handleClose}
-                                >
-                                    <MenuItem onClick={this.handleClose}>Profile</MenuItem>
-                                    <MenuItem onClick={this.handleClose}>My account</MenuItem>
-                                </Menu>
-                            </div>
-                        )}
-                        <FormControlLabel
-                            control={
-                                <Switch checked={auth} onChange={this.handleChange} aria-label="LoginSwitch"/>
-                            }
-                            label={auth ? 'Logout' : 'Login'}
-                        />
-                    </Toolbar>
-                </AppBar>
-            </div>
+            <AppBar position="static" className={classes.appBar}>
+                <Toolbar>
+                    <Link to="/"><Logo/></Link>
+                    <Typography type="title" color="inherit" className={classes.flex}>
+                        BibGallery
+                    </Typography>
+                    {this.props.auth && (
+                        <div>
+                            <IconButton
+                                aria-owns={open ? 'menu-appbar' : null}
+                                aria-haspopup="true"
+                                onClick={this.handleMenu}
+                                color="inherit"
+                            >
+                                <AccountCircle/>
+                            </IconButton>
+                            <Menu
+                                id="menu-appbar"
+                                anchorEl={anchorEl}
+                                anchorOrigin={{
+                                    vertical: 'top',
+                                    horizontal: 'right',
+                                }}
+                                transformOrigin={{
+                                    vertical: 'top',
+                                    horizontal: 'right',
+                                }}
+                                open={open}
+                                onClose={this.handleClose}
+                            >
+                                <MenuItem onClick={this.handleClose}>Profile</MenuItem>
+                                <MenuItem onClick={this.handleClose}>My account</MenuItem>
+                            </Menu>
+                        </div>
+                    )}
+                    <FormControlLabel
+                        control={
+                            <Switch checked={this.props.auth} onChange={this.handleChange} aria-label="LoginSwitch"/>
+                        }
+                        label={this.props.auth ? 'Logout' : 'Login'}
+                    />
+                </Toolbar>
+            </AppBar>
         );
     }
 }
@@ -105,4 +105,17 @@ Header.propTypes = {
     classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(Header);
+const mapStateToProps = state => {
+    return {
+        auth: state.auth
+    };
+};
+
+const mapDispatchToProps = dispatch => {
+    return {
+        onLogin: () => dispatch({type: 'LOGIN'}),
+        onLogout: () => dispatch({type: 'LOGOUT'})
+    };
+};
+
+export default withStyles(styles)(connect(mapStateToProps, mapDispatchToProps)(Header));
