--- src/hoc/ReactAux.js
@@ -1,2 +1,0 @@
-const reactAux=(props)=>props.children;
-export default reactAux;
