+++ src/hoc/ReactAux/ReactAux.js
@@ -1,0 +1,2 @@
+const reactAux=(props)=>props.children;
+export default reactAux;
