+++ src/index.js
@@ -1,8 +1,16 @@
+// React and Redux imports:
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {createStore} from 'redux';
+import {Provider} from 'react-redux'
+
+// Others:
 import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
+import reducer from './store/reducer'
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const store = createStore(reducer);
+
+ReactDOM.render(<Provider store={store}><App/></Provider>, document.getElementById('root'));
 registerServiceWorker();
