+++ src/serverConfig.js
@@ -1,0 +1,4 @@
+module.exports = {
+    backendUrl: 'http://localhost:3001/'
+
+};
