+++ src/store/reducer.js
@@ -1,0 +1,22 @@
+const initialState = {
+    auth: false
+}
+
+const reducer = (state = initialState, action) => {
+    switch (action.type) {
+        case 'LOGIN':
+            return {
+                ...state,
+                auth: true
+            };
+        case 'LOGOUT':
+            return {
+                ...state,
+                auth: false
+            };
+    }
+
+    return state;
+};
+
+export default reducer;
